+++ Documentation/Changelog.md
@@ -4,6 +4,7 @@
 
 * Runtime Configuration of the Cache
 * User Specified Syslog Facility and Level for Authentication Errors
+* `config reload` removed from MaxAdmin (was deprecated in 2.2)
 
 For more details, please refer to: